Move the constantIdProducer closer to presentation logic
