+++ Architecture/TodoApp/android/mobile/src/main/java/com/novoda/todoapp/task/edit/EditTaskActivity.java
@@ -9,7 +9,6 @@ import com.novoda.todoapp.navigation.AndroidNavigator;
 import com.novoda.todoapp.task.data.model.Id;
 import com.novoda.todoapp.task.edit.displayer.EditTaskDisplayer;
 import com.novoda.todoapp.task.edit.presenter.EditTaskPresenter;
-import com.novoda.todoapp.task.presenter.IdProducer;
 
 public class EditTaskActivity extends AppCompatActivity {
 
@@ -24,12 +23,7 @@ public class EditTaskActivity extends AppCompatActivity {
                 TodoApplication.TASKS_SERVICE,
                 ((EditTaskDisplayer) findViewById(R.id.content)),
                 new AndroidNavigator(this),
-                new IdProducer() {
-                    @Override
-                    public Id produce() {
-                        return taskId;
-                    }
-                }
+                taskId
         );
     }
